fixing
Hi, in NB.cpp on line 149 I find int ready = MODEM.ready();, this calls MODEM.poll() internally to update the internal state, then reads and returns the state. What is the reason for the extra MODEM.poll();on line 155?

I have commented it out in my copy of the library. The code runs fine without it, and it is not present in the parallel library for MKR 1400 GSM.

I recommend this change hope it helps :)