Delay of 60 in USOWR
From the UBlox AT Commands example App note:

Request to write 100 data bytes into socket #0. Wait for "@" symbol indicating the data prompt is now open (AT commands are not allowed in data prompt).
After the @ prompt reception, wait for a minimum of 50 ms before sending data.

At the moment, there is no wait in the lib. I added a patch that sends the USOWR command and then wait blindly for 60 ms. I experienced troubles while waiting for the '@' prompt, so i just added a stupid delay(60) to fullfill the specs better than now. It's just a patch that can be certainly done better...